agentflow/server/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
crewai>=0.102.0
crewai-tools>=0.17.0
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
